別ファイルに関数移動
